% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAMU-package.R
\docType{package}
\name{GSAMU-package}
\alias{GSAMU-package}
\title{GSAMU: Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding: non-Gaussian and time-to-event outcomes}
\description{
GSAMU is a package for accommodating statistical models involving Gaussian, non-Gaussian, and time-to-event outcome variables and multiple exposures, proving indispensable in the context of unmeasured confounding.
}
\references{
Lee S, Jeong B, Lee D, Lee W (2024+):
Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding: non-Gaussian and time-to-event outcomes
\emph{xxx}. DOI: xxx.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sites.google.com/view/lwj221/research/softwares}
  \item \url{https://github.com/seungjae2525/GSAMU}
  \item Report bugs at \url{https://github.com/seungjae2525/GSAMU/issues}
}

}
\author{
\strong{Maintainer}: Seungjae Lee \email{seungjae2525@gmail.com}

Authors:
\itemize{
  \item Boram Jeong
  \item Donghwan Lee
  \item Woojoo Lee
}

}
