% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GSAMU.R
\name{print.GSAMU}
\alias{print.GSAMU}
\title{Print for \code{GSAMU} objects}
\usage{
\method{print}{GSAMU}(x, digits = max(1L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object for class \code{GSAMU}.}

\item{digits}{Print digits. Default: max(1L, getOption("digits") - 3L).}

\item{...}{Further arguments (currently not used).}
}
\description{
Print the sensitivity analysis results for object of class \code{GSAMU}.
}
\details{
Print the sensitivity analysis results for object (\code{GSAMU}) of class \code{GSAMU}.
}
\examples{
############################################
## Real data analysis for binary outcome ##
############################################
### Log transformed of exposures. continuous variables are standardized
library(readr)
ul=paste0("https://static-content.springer.com/esm/art\%3A10.1186\%2Fs12940-020-00644-4/",
          "MediaObjects/12940_2020_644_MOESM2_ESM.csv")
data_orig <- as.data.frame(read_csv(url(ul))); data_orig <- na.omit(data_orig)

### Data cleaning
## Make binary outcome
data_orig$triglyceride_b <- ifelse(data_orig$triglyceride >= 150, 1, 0)
data_orig$triglyceride <- NULL

### Environmental factors
## All exposure variables, except for retinol, were log-transformed,
# Retinyl-palmitate
data_orig$a3.Retinyl.palmitate <- log(data_orig$a3.Retinyl.palmitate)
# Retinol
summary(data_orig$a5.Retinol)
# trans.b.carotene
data_orig$a1.trans.b.carotene <- log(data_orig$a1.trans.b.carotene)
# a-Tocopherol
data_orig$a7.a.Tocopherol <- log(data_orig$a7.a.Tocopherol)
# g-tocopherol
data_orig$a6.g.tocopherol <- log(data_orig$a6.g.tocopherol)

### Covariates
# Age
summary(data_orig$age)
# Sex
data_orig$sex <- ifelse(data_orig$sex == 2, 1, 0)
# Race [Race/Ethnicity (1: Non-Hispanic white; 2: Non-Hispanic black; 3: Mexican American;
#                       4: Other race - Including multi-racial; 5: Other Hispanic)]
data_orig$race <- ifelse(data_orig$rac == 1, 1, 0)

## using only needed variables
usedata <- data_orig[,c("triglyceride_b", "sex", "race", "age",
                        "a3.Retinyl.palmitate", "a5.Retinol", "a1.trans.b.carotene",
                        "a7.a.Tocopherol", "a6.g.tocopherol")]
## Rename the varoables
names(usedata) <- c("triglyceride_b", "sex", "race", "age",
                    "Retinyl palmitate", "Retinol", "trans-beta-carotene",
                    "alpha-Tocopherol", "gamma-Tocopherol")
data_r4 <- usedata

## glm fitting (working regression model)
fit.glm <- glm(triglyceride_b~., family=binomial(link="logit"), data=data_r4)
summary(fit.glm)
round(cbind("Est"=exp(coef(fit.glm)), exp(confint(fit.glm)),
      "P value"=summary(fit.glm)$coefficients[,4]), 3)

##
beta <- glm(triglyceride_b~., family=binomial(link="logit"), data=data_r4)$coefficient

##
bound <- c(## upper bound
  # sex   race   age
  0.32, 0.5, 0.29,
  # Retinyl palmitate    Retinol  trans-beta-carotene
  rep(0.15, 3),
  # alpha-Tocopherol  gamma-Tocopherol
  rep(0.10, 2),

  ## lower bound
  # sex   race   age
  0.0, 0.0, 0.0,
  # Retinyl palmitate    Retinol  trans-beta-carotene
  rep(-0.19, 3),
  # alpha-Tocopherol  gamma-Tocopherol
  rep(-0.19, 2))

## Sensitivity analysis
binary.re0 <- GSAMU(data=data_r4, outcome="triglyceride_b", outcome.type="binary", link="logit",
                    hazard.model=NULL,
                    confounder=c("sex", "race", "age"),
                    exposure=c("Retinyl palmitate", "Retinol", "trans-beta-carotene",
                               "alpha-Tocopherol", "gamma-Tocopherol"),
                    delta=c(0.11, 0.22, 0.33, 0.44), bound=bound,
                    bootsCI=FALSE, B=1000, seed=231111, verbose=TRUE)
print(binary.re0)

\dontrun{
## Sensitivity analysis with bootstrap percentile confidence interval
binary.re1 <- GSAMU(data=data_r4, outcome="triglyceride_b", outcome.type="binary", link="logit",
                    hazard.model=NULL,
                    confounder=c("sex", "race", "age"),
                    exposure=c("Retinyl palmitate", "Retinol", "trans-beta-carotene",
                               "alpha-Tocopherol", "gamma-Tocopherol"),
                    delta=c(0.11, 0.22, 0.33, 0.44), bound=bound,
                    bootsCI=TRUE, B=1000, seed=231111, verbose=TRUE)
print(binary.re1)
}

}
\seealso{
\code{\link[GSAMU]{GSAMU}}
}
\keyword{Print}
