% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAMU.R
\name{GSAMU}
\alias{GSAMU}
\title{Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding: non-Gaussian and time-to-event outcomes}
\usage{
GSAMU(
  data,
  outcome,
  outcome.type,
  link = NULL,
  hazard.model = NULL,
  confounder,
  exposure,
  delta.range,
  delta.diff,
  bound,
  bootsCI = TRUE,
  B = 1000,
  seed = 231111,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame in which contains outcome, measured confounders, exposures.}

\item{outcome}{The name of variable for outcome. For count and binary outcomes, a single character for outcome is specified. For time-to-event outcome, two characters corresponding to survival time and status are specified (i.e., c("survivaltime", "status")).}

\item{outcome.type}{The type of outcome variable. Possible values are "count" (for count outcome), "binary" (for binary outcome), or "timetoevent" (for time-to-event outcome).}

\item{link}{The specification for the model link function. Possible values are "log" (for count outcome), "probit" (for binary outcome), or "logit" (for binary outcome). Not used for outcome.type = "timetoevent".}

\item{hazard.model}{The specification for the hazard model. Possible values are "coxph" (for Cox PH model) or "ah" (for additive hazard model). Not used for outcome.type = "binary" or "count".}

\item{confounder}{The vector of variable names for confounders. See Examples.}

\item{exposure}{The vector of variable names for exposures. See Examples.}

\item{delta.range}{The range of \eqn{(\delta_{min}, \delta_{max})}. See Examples.}

\item{delta.diff}{The increment of the sequence of \eqn{\delta}. Default: 0.1.}

\item{bound}{The range of \eqn{(\phi_{1}, \ldots, \phi_{k})} and \eqn{(\rho_{1}, \ldots, \rho_{p})}. The order of inputs is (\eqn{\phi_{1,max}, \ldots, \phi_{k,max}, } \eqn{\rho_{1,max}, \ldots, \rho_{p,max}, \phi_{1,min}, \ldots, \phi_{k,min}, \rho_{1,min}, \ldots, \rho_{p,min}}). See Examples.}

\item{bootsCI}{Logical value: if TRUE, The process to obtain the bootstrap confidence interval is carried out. Default: TRUE.}

\item{B}{The number of bootstrap replicates. Not used for bootsCI=FALSE. Default: 1000}

\item{seed}{The value of seed number when bootstrap works. Not used for bootsCI=FALSE. Default: 231111.}

\item{verbose}{Logical value: if TRUE, bootstrap process is output every 100 times. Not used for bootsCI=FALSE. Default: TRUE.}
}
\value{
An object of class \code{condSens}. The object is a data.frame with the following components:
\item{label}{Exposures}
\item{delta}{The values of \eqn{\delta}}
\item{coef}{The coefficients of exposures}
\item{Lower.SI}{The lower sensitivity interval of conditional single- or joint-exposure effect}
\item{Upper.SI}{The upper sensitivity interval of conditional single- or joint-exposure effect}
\item{Lower.CI}{The lower confindece interval of sensitivity interval. Results are displayed only when bootstrap is performed.}
\item{Upper.CI}{The upper confindece interval of sensitivity interval. Results are displayed only when bootstrap is performed.}

The results for the \code{GSAMU} are printed with the \code{\link{print.GSAMU}} function.
To generate the plot of results for the \code{GSAMU}, use the \code{\link{autoplot.GSAMU}} function.
}
\description{
\code{GSAMU()} is the main function of GSAMU package. It performs a novel sensitivity analysis for non-Gaussian and time-to-event outcomes with multiple exposures. It also enables researchers to assess the vulnerability of the conditional exposure effects to the degree of unmeasured confounding for various outcome types.
}
\details{
See Lee et al. (2024+) for details.
}
\examples{
############################################
## Real data analysis for binary outcome ##
############################################
### Log transformed of exposures. continuous variables are standardized
library(readr)
ul=paste0("https://static-content.springer.com/esm/art\%3A10.1186\%2Fs12940-020-00644-4/",
          "MediaObjects/12940_2020_644_MOESM2_ESM.csv")
data_orig <- as.data.frame(read_csv(url(ul))); data_orig <- na.omit(data_orig)

### Data cleaning
## Make binary outcome
data_orig$triglyceride_b <- ifelse(data_orig$triglyceride >= 150, 1, 0)
data_orig$triglyceride <- NULL

### Environmental factors
## All exposure variables, except for retinol, were log-transformed,
# Retinyl-palmitate
data_orig$a3.Retinyl.palmitate <- log(data_orig$a3.Retinyl.palmitate)
# Retinol
summary(data_orig$a5.Retinol)
# trans.b.carotene
data_orig$a1.trans.b.carotene <- log(data_orig$a1.trans.b.carotene)
# a-Tocopherol
data_orig$a7.a.Tocopherol <- log(data_orig$a7.a.Tocopherol)
# g-tocopherol
data_orig$a6.g.tocopherol <- log(data_orig$a6.g.tocopherol)

### Covariates
# Age
summary(data_orig$age)
# Sex
data_orig$sex <- ifelse(data_orig$sex == 2, 1, 0)
# Race [Race/Ethnicity (1: Non-Hispanic white; 2: Non-Hispanic black; 3: Mexican American;
#                       4: Other race - Including multi-racial; 5: Other Hispanic)]
data_orig$race <- ifelse(data_orig$rac == 1, 1, 0)

## using only needed variables
usedata <- data_orig[,c("triglyceride_b", "sex", "race", "age",
                        "a3.Retinyl.palmitate", "a5.Retinol", "a1.trans.b.carotene",
                        "a7.a.Tocopherol", "a6.g.tocopherol")]
## Rename the varoables
names(usedata) <- c("triglyceride_b", "sex", "race", "age",
                    "Retinyl palmitate", "Retinol", "trans-beta-carotene",
                    "alpha-Tocopherol", "gamma-Tocopherol")
data_r4 <- usedata

## glm fitting (working regression model)
fit.glm <- glm(triglyceride_b~., family=binomial(link="logit"), data=data_r4)
summary(fit.glm)
round(cbind("Est"=exp(coef(fit.glm)), exp(confint(fit.glm)),
      "P value"=summary(fit.glm)$coefficients[,4]), 3)

##
beta <- glm(triglyceride_b~., family=binomial(link="logit"), data=data_r4)$coefficient

##
bound <- c(## upper bound
  # sex   race   age
  0.32, 0.5, 0.29,
  # Retinyl palmitate    Retinol  trans-beta-carotene
  rep(0.15, 3),
  # alpha-Tocopherol  gamma-Tocopherol
  rep(0.10, 2),

  ## lower bound
  # sex   race   age
  0.0, 0.0, 0.0,
  # Retinyl palmitate    Retinol  trans-beta-carotene
  rep(-0.19, 3),
  # alpha-Tocopherol  gamma-Tocopherol
  rep(-0.19, 2))

## Sensitivity analysis
binary.re0 <- GSAMU(data=data_r4, outcome="triglyceride_b", outcome.type="binary", link="logit",
                    hazard.model=NULL,
                    confounder=c("sex", "race", "age"),
                    exposure=c("Retinyl palmitate", "Retinol", "trans-beta-carotene",
                               "alpha-Tocopherol", "gamma-Tocopherol"),
                    delta.range=c(0.11, 0.44), delta.diff=0.11, bound=bound,
                    bootsCI=FALSE, B=1000, seed=231111, verbose=TRUE)
binary.re0$result

\dontrun{
## Sensitivity analysis with bootstrap percentile confidence interval
binary.re1 <- GSAMU(data=data_r4, outcome="triglyceride_b", outcome.type="binary", link="logit",
                    hazard.model=NULL,
                    confounder=c("sex", "race", "age"),
                    exposure=c("Retinyl palmitate", "Retinol", "trans-beta-carotene",
                               "alpha-Tocopherol", "gamma-Tocopherol"),
                    delta.range=c(0.11, 0.44), delta.diff=0.11, bound=bound,
                    bootsCI=TRUE, B=1000, seed=231111, verbose=TRUE)
binary.re1$result
}

}
\references{
Lee S, Jeong B, Lee D, Lee W (2024+):
Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding: non-Gaussian and time-to-event outcomes
\emph{xxx}. DOI: xxx.
}
\seealso{
\code{\link[GSAMU]{print.GSAMU}}, \code{\link[GSAMU]{autoplot.GSAMU}}
}
\keyword{Methods}
